move QueryParamEncodable to its own file and write tests
