+++ ConsentViewController/Classes/SourcePointClient/ConsentStatusMetadata.swift
@@ -7,21 +7,7 @@
 
 import Foundation
 
-protocol QueryParamEncodable: Encodable {
-    func stringified() -> String
-}
-
-extension QueryParamEncodable {
-    func stringified() -> String {
-        guard let data = try? JSONEncoder().encode(self),
-              let metadataString = String(data: data, encoding: .utf8) else {
-                  return "{}"
-              }
-
-        return metadataString
-    }
-}
-
+/// Data Transfer Class used to encapsulate the meta data of `/consent-status` network call
 struct ConsentStatusMetaData: QueryParamEncodable {
     struct Campaign: Encodable {
         let hasLocalData, applies: Bool
